refactor: extract register step 1 to file
