+++ src/components/new-user.ts
@@ -1,17 +1,16 @@
 import { ethers } from "ethers";
-import { appState } from "../main";
-import { wirexPayChain, wirexPayChainTestnet } from "../shared/wirex-pay-chain";
-import { AccessToken } from "../../functions/shared";
+import { backendBaseUrl } from "../constants";
+import { registerOnChain } from "./register/on-chain-register";
 import { showToast } from "./toaster";
 
 // Step states
-enum RegistrationStep {
+export enum RegistrationStep {
   INITIAL,
   ON_CHAIN_REGISTERED,
   API_REGISTERED,
 }
 
-const backendBaseUrl = "";
+const totalRegistrationSteps = 2;
 
 let currentStep = RegistrationStep.INITIAL;
 
@@ -67,8 +66,6 @@ export function newUser(): string {
   `;
 }
 
-const totalRegistrationSteps = 2;
-
 export function handleNewUserEvents() {
   // Step 1: On-chain registration
   document.getElementById("register")?.addEventListener("click", (event) => {
@@ -81,6 +78,7 @@ export function handleNewUserEvents() {
       } catch (error) {
         console.error(error);
       }
+      updateRegistrationUi();
       button.style.pointerEvents = "auto"; // Re-enable clicks
     })().catch(console.error);
   });
@@ -91,32 +89,6 @@ export function handleNewUserEvents() {
     showToast({ message: `Step 2/${totalRegistrationSteps}: Register on wirex.` });
     registerOnApp().catch(console.error);
   });
-}
-
-async function registerOnChain(button: HTMLAnchorElement) {
-  try {
-    const authData = await authenticateUser();
-    if (!authData) return;
-
-    const wirexRegisterContract = await setupNetworkAndContract(authData);
-
-    if (!window.ethereum) {
-      alert("Please install a Web3 wallet like MetaMask to continue.");
-      return;
-    }
-
-    const provider = new ethers.providers.Web3Provider(window.ethereum);
-    const isRegistered = await checkUserRegistration(provider, wirexRegisterContract);
-
-    if (isRegistered) {
-      console.log("User already registered on-chain, proceeding to step 2");
-      updateRegistrationUi();
-    } else {
-      await registerNewUser(provider, wirexRegisterContract);
-    }
-  } finally {
-    button.style.pointerEvents = "auto"; // Re-enable clicks
-  }
 }
 
 async function registerOnApp() {
@@ -157,66 +129,8 @@ async function registerOnApp() {
   }
 }
 
-// Check if user is already registered on-chain
-async function checkUserRegistration(provider: ethers.providers.Web3Provider, contractAddress: string): Promise<boolean> {
-  try {
-    const signer = provider.getSigner();
-    const userAddress = await signer.getAddress();
-
-    // ABI for checking if user is registered
-    const checkRegistrationAbi = [
-      {
-        inputs: [
-          {
-            internalType: "address",
-            name: "owner",
-            type: "address",
-          },
-        ],
-        name: "getAccount",
-        outputs: [
-          {
-            components: [
-              {
-                internalType: "enum IAccounts.AccountStatus",
-                name: "status",
-                type: "uint8",
-              },
-              {
-                internalType: "enum IAccounts.AccountVerificationStatus",
-                name: "verificationStatus",
-                type: "uint8",
-              },
-            ],
-            internalType: "struct IAccounts.Account",
-            name: "",
-            type: "tuple",
-          },
-        ],
-        stateMutability: "view",
-        type: "function",
-      },
-    ];
-
-    const contract = new ethers.Contract(contractAddress, checkRegistrationAbi, provider);
-    const accountInfo = await contract.getAccount(userAddress);
-
-    // AccountStatus enum: 0 = NOT_CREATED, 1 = CREATED, 2 = SUSPENDED
-    // We consider the account registered if status is CREATED (1) or SUSPENDED (2)
-    const accountStatus = accountInfo.status;
-    console.log(`Account status: ${accountStatus}`);
-    return accountStatus > 0; // If status > 0, account exists
-  } catch (error) {
-    console.error("Error checking registration:", error);
-    return false;
-  }
-}
-
-// Register user with API
 async function registerUserWithApi(email: string, userAddress: string): Promise<boolean> {
   try {
-    // API endpoint from WirexPayChain partner documentation
-
     const response = await fetch(`${backendBaseUrl}/register`, {
       method: "POST",
       headers: {
@@ -246,51 +160,10 @@ async function registerUserWithApi(email: string, userAddress: string): Promise<
   }
 }
 
-async function authenticateUser() {
-  const authUrl = `${backendBaseUrl}/auth`;
-  const authResponse = await fetch(authUrl, { method: "GET" });
-  const responseJson = await authResponse.json();
-
-  if (authResponse.status !== 200) {
-    alert(`Error authenticating: ${responseJson}`);
-    console.error("Error authenticating: ", responseJson);
-    return null;
-  }
-  return responseJson;
-}
-
-async function setupNetworkAndContract(responseJson: AccessToken) {
-  const isSandbox = responseJson.isSandbox;
-  await appState.switchNetwork(isSandbox ? wirexPayChainTestnet : wirexPayChain);
-
-  return isSandbox ? "0x3fe04562Fc28b4152F24A41E8A8c3899E6B8c433" : "0x2766F66E572C94a4cbc57f4d5bd2aD71900edF30";
-}
-
-function updateRegistrationUi() {
+export function updateRegistrationUi() {
   const step1 = document.getElementById("step-1");
   const step2 = document.getElementById("step-2");
   if (step1) step1.style.display = "none";
   if (step2) step2.style.display = "block";
   currentStep = RegistrationStep.ON_CHAIN_REGISTERED;
-}
-
-async function registerNewUser(provider: ethers.providers.JsonRpcProvider, contractAddress: string) {
-  const signer = provider.getSigner();
-  const wirexContractAbi = [
-    {
-      type: "function",
-      name: "createAccount",
-      constant: false,
-      payable: false,
-      inputs: [],
-      outputs: [],
-    },
-  ];
-
-  const wirexContract = new ethers.Contract(contractAddress, wirexContractAbi, signer);
-  const tx = await wirexContract.createAccount();
-  await tx.wait();
-
-  alert("Successfully registered on-chain! Please complete step 2.");
-  updateRegistrationUi();
 }
